% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLINE.R
\name{scLINE}
\alias{scLINE}
\title{Integrated networks embedding for scRNA-seq data}
\usage{
scLINE(mat, network, L, K = 5, T, rho = 0.025, Random_weight)
}
\arguments{
\item{mat}{A matrix of scRNA-seq data in which rows represent gene columns representing cells}

\item{network}{A list of gene networks with 3 columns: geneID, geneID, score}

\item{L}{Dim of the low-dimensional representations}

\item{K}{Number of negative samples, the default is 5}

\item{T}{Number of iterations}

\item{rho}{The initial learning rate, the default is 0.025}

\item{Random_weight}{True of False,Whether to add random weights to the network model}
}
\value{
the low-dimensional matrix of input data
}
\description{
A tool of dimension reduction for scRNA-seq data combined with multiple gene networks embedding model
}
\examples{
load(system.file("data","Usoskin.Rdata",package = "scLINE"))
load(system.file("data","ppi.Rdata",package = "scLINE"))
load(system.file("data","humannet.Rdata",package = "scLINE"))
exp_mat<-Usoskin$rawdata
gene_network<-list(ppi = ppi,humannet = humannet)
lowdim_list<-scLINE(exp_mat, gene_network, L = 20, K = 5, T = 1e7, rho = 0.025, Random_weight = TRUE)
}
