% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualization}
\usage{
visualize(data, label)
}
\arguments{
\item{data}{a matrix in which each row represents a cell}

\item{label}{true labels of the cell corresponding to  data}
}
\value{
Visualization image of the data
}
\description{
Visualize the low-dimensional vectors obtained by scLINE in 2D
}
\examples{
visualize(low_mat$cell_low,Usoskin$label)
}
